refactor(sn): rename `routing::peer` to `routing::peers`
This is a more general name for the module, making it appropriate to put
things besides the `Peer` structs (e.g. `ElderCandidates`...).
