+++ sn/src/routing/relocation.rs
@@ -248,7 +248,7 @@ mod tests {
     use crate::messaging::SectionAuthorityProvider;
     use crate::routing::{
         dkg::test_utils::section_signed, network_knowledge::NodeStateUtils,
-        peer::test_utils::arbitrary_unique_peers, routing_api::tests::SecretKeySet,
+        peers::test_utils::arbitrary_unique_peers, routing_api::tests::SecretKeySet,
         SectionAuthorityProviderUtils, ELDER_SIZE, MIN_AGE,
     };
     use assert_matches::assert_matches;
